voxa-emotion-engine/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
deepface
librosa
transformers
torch
Pillow